Refresh AI core chat terminology and messaging
